Ya anda el login

+++ FrbaCommerce/Entidades/Ent_Usuario.cs
@@ -9,7 +9,7 @@ namespace FrbaCommerce.Entidades
     {
         public String Usuario { get; set; }
         public String Contraseña { get; set; }
-        public Int16 Rol { get; set; }
+        public Decimal Rol { get; set; }
 
     }
 }
